move default channel to cable ready module
